[libromdata] Fix MSVC /W3 /WX warnings in the 32-bit (i386) Windows build.
Mostly related to converting `off64_t` (file sizes) to `size_t`.

J2MEPrivate::loadFileFromZip():
- Cast the default max_size parameter to `size_t`.
- Cast file_info.uncompressed_size to `size_t` as well. This should be
  optimized out in 64-bit builds.

ELF:
- Need to cast d_tag values to `size_t` before using it with operator[].

CdiReader, GdiReader:
- Cast std::numeric_limits<>::max() to `size_t` before comparing it
  with idx.

+++ src/libromdata/disc/CdiReader.cpp
@@ -339,8 +339,8 @@ int CdiReaderPrivate::parseCdiFile(void)
 			if (lengthFields.mode != 0) {
 				// Save the track information.
 				const size_t idx = blockRanges.size();
-				assert(idx < std::numeric_limits<int8_t>::max());
-				if (idx >= std::numeric_limits<int8_t>::max()) {
+				assert(idx < static_cast<size_t>(std::numeric_limits<int8_t>::max()));
+				if (idx >= static_cast<size_t>(std::numeric_limits<int8_t>::max())) {
 					// Too many tracks. (More than 127???)
 					return -ENOMEM;
 				}
