+++ src/libromdata/disc/GdiReader.cpp
@@ -219,8 +219,8 @@ int GdiReaderPrivate::parseGdiFile(char *gdibuf)
 
 		// Save the track information.
 		const size_t idx = blockRanges.size();
-		assert(idx < std::numeric_limits<int8_t>::max());
-		if (idx >= std::numeric_limits<int8_t>::max()) {
+		assert(idx < static_cast<size_t>(std::numeric_limits<int8_t>::max()));
+		if (idx >= static_cast<size_t>(std::numeric_limits<int8_t>::max())) {
 			// Too many tracks. (More than 127???)
 			return -ENOMEM;
 		}
